refactoring to be able to anonymize text block
The library was initially designed for the clean up of nested data structures.
We now also need to clean up regular text blocks.

The library now provides two differents functions depending to deal with the two situation:

- `anonymizor.anonymize_struct(example)`: to clean up a dictionary.
- `anonymizor.anonymize_text_block(some_text)`: to clean up a text block.

Also, the library was adjusted in order to avoid the use of random numbers. This in
order to get deterministic results that suite the tests better.