hide_secrets: better management of the quotes
This PR introduces a tiny parser that can parse the YAML-like structure
and identify the fields that need to by anonymized.
The parser is tolerant to YAML format errors and can also parse
similar format like ini or toml.

Internally, the parser builds a tree. Navigation is possible:
- through the original text order, from left to right
- or right to left
- or through the different level of quoted strings.

The change will avoid the following issues:
- a trailing quote is remove with the password and we break the YAML
- quotes are insered around the Jinja2 placeholder in a block that is
  already quoted.